+++ src/data_struct/feature_elem.hpp
@@ -13,7 +13,7 @@ public:
     FeatureElem(std::istringstream &line_conv, std::vector<float> &count_vect, std::ofstream &idx_file, const TabHeader &tab_header);
 
     const void EvalFeatureElem(const std::unique_ptr<Scorer> &scorer);
-    void ScaleValue(float fact, float lower_lim, float upper_lim);
+    void ScaleValue(double fact, double lower_lim, double upper_lim);
     const std::vector<float> &GetNormCondiMeans() const;
 
 private:
